+++ manta-pay/src/crypto/hash/poseidon/mod.rs
@@ -23,7 +23,7 @@ use alloc::vec::Vec;
 use core::{fmt::Debug, hash::Hash, iter, mem};
 use manta_crypto::{
     hash::ArrayHashFunction,
-    rand::{CryptoRng, RngCore, Sample},
+    rand::{RngCore, Sample},
 };
 use manta_util::{
     codec::{Decode, DecodeError, Encode, Read, Write},
@@ -344,11 +344,10 @@ where
     S: Specification<COM>,
     S::ParameterField: Field + FieldGeneration + PartialEq + Sample<D>,
 {
-    /// Samples random Poseidon parameters.
     #[inline]
     fn sample<R>(distribution: D, rng: &mut R) -> Self
     where
-        R: CryptoRng + RngCore + ?Sized,
+        R: RngCore + ?Sized,
     {
         let _ = (distribution, rng);
         Self {